Rename the namespace to Cake\Queue
After going through the RFC the bulk of the support for the rename was
to keep something simple and not cute. Furthermore a few folks requested
that we didn't use the top level `Queue` namespace. Our historical
convention on this has been mixed.

Most of our plugins do not have the Cake prefix (Authorization,
Authentication, Bake, DebugKit, Migrations), however ElasticSearch does
as it overlaps another library that has a similar name. I think this
library is in a similar position with Enqueue.